Mag Heading must be an integer
git-svn-id: file:///home/svn/DVA2006@18816 1e6d2d28-e1fb-0310-b8ab-81908d8c6b6a

+++ src/java/org/deltava/service/navdata/XMLFormatter.java
@@ -119,7 +119,7 @@ class XMLFormatter {
 			rwe.setAttribute("lat", df.format(r.getLatitude()));
 			rwe.setAttribute("lng", df.format(r.getLongitude()));			
 			rwe.setAttribute("width", String.valueOf(r.getWidth()));
-			rwe.setAttribute("hdg", String.valueOf(GeoUtils.normalize(r.getHeading() + r.getMagVar())));
+			rwe.setAttribute("hdg", String.valueOf((int)GeoUtils.normalize(r.getHeading() + r.getMagVar())));
 			rwe.setAttribute("trueHdg", String.valueOf(r.getHeading()));
 			rwe.setAttribute("length", String.valueOf(r.getLength()));
 			rwe.setAttribute("sim", r.getSimulator().name());
